django_app/requirement.txt
# Django Core
Django==4.2.7
django-environ==0.11.2

# Authentication and Security
django-auth-keycloak==0.3.0
python-jose==3.3.0
social-auth-core==4.4.2
social-auth-app-django==5.2.0

# API and Requests
requests==2.31.0
djangorestframework==3.14.0
django-cors-headers==4.2.0

# Canvas LMS Integration
canvasapi==3.2.0

# Database Connectors
psycopg2-binary==2.9.9
django-postgres-extensions==0.5.0

# Caching
django-redis==5.3.0
redis==5.0.1

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Logging and Monitoring
sentry-sdk==1.37.1
django-health-check==3.17.0

# JWT Token Handling
PyJWT==2.8.0
cryptography==41.0.5

# OAuth2 and OpenID Connect
oauthlib==3.2.2
django-oauth-toolkit==2.3.0

# File Handling
django-storages==1.14.2
boto3==1.28.74

# Testing
pytest==7.4.3
pytest-django==4.5.2
coverage==7.3.2
model-bakery==1.16.0

# Performance and Optimization
django-debug-toolbar==4.2.0
gunicorn==21.2.0
whitenoise==6.5.0

# Security Extras
django-axes==6.1.1
django-defender==0.9.6

# Internationalization
Babel==2.13.1
pytz==2023.3.post1

# Documentation
drf-yasg==1.21.7

# Additional Utilities
python-slugify==8.0.1
Markdown==3.5.1
packaging==23.2

# Development Tools
ipython==8.17.2
black==23.10.1
flake8==6.1.0
isort==5.12.0